DocDB compaction filter
Summary:
Adding a DocDB compaction filter that garbage-collects key/value pairs in RocksDB that are not visible at a specified "history cut-off timestamp". We are using the `CompactionFilter` / `CompactionFilterFactory` framework available in RocksDB to run custom logic on compactions. `CompactionFilterFactory` allows us to parameterize a `DocDBCompactionFilter` with a history cutoff timestamp at compaction start time and not worry about its thread safety, because a new `CompactionFilter` is used for each compaction.

Also adding extensive randomized tests for reading at old timestamps and verifying that the state of the DB is as expected.

Currently the retention policy is simply to retain history for a fixed time interval, e.g. 60 seconds or 10 seconds.

Future work:
- We should start tracking the read point and make sure we don't collect history at timestamps that data is still being read at, but that is not part of this revision.
- Only garbage-collecting old versions on "full" (or "major") compactions. We should be able to do the same on other types of compactions in the absence of cross-tablet transactions, but that requires separate tests and is thus left to a further revision.

Test Plan: Jenkins

Reviewers: kannan, akashnil

Reviewed By: kannan, akashnil

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D810

+++ src/yb/client/client_samples-test.sh
@@ -23,6 +23,8 @@
 
 set -e -u -o pipefail
 
+keep_library_dir=""
+
 # Clean up after the test. Must be idempotent.
 cleanup() {
   if [ -n "$TS_PID" ]; then
@@ -36,8 +38,10 @@ cleanup() {
   if [ -n "$BASE_DIR" -a -d "$BASE_DIR" ]; then
       rm -rf "$BASE_DIR"
   fi
-  if [ -n "$LIBRARY_DIR" -a -d "$LIBRARY_DIR" ]; then
-      rm -rf "$LIBRARY_DIR"
+  if [[ -z ${keep_library_dir} ]]; then
+    if [ -n "$LIBRARY_DIR" -a -d "$LIBRARY_DIR" ]; then
+        rm -rf "$LIBRARY_DIR"
+    fi
   fi
 }
 trap cleanup EXIT
@@ -132,5 +136,37 @@ TS_PID=$!
 # Let them run for a bit.
 sleep 5
 
+set +e
+
+ulimit -c unlimited  # Ignore errors in setting the core dump size limit.
+
 # Run the samples.
 $SAMPLES_DIR/sample $LOCALHOST_IP
+exit_code=$?
+
+set +x
+
+if [[ $exit_code -ne 0 ]]; then
+  echo "Current directory: $PWD"
+  echo "LIBRARY_DIR: $LIBRARY_DIR"
+  echo "TEST_TMPDIR: $TEST_TMPDIR"
+fi
+
+core_found=false
+for core_path in core.*; do
+  if [[ -f $core_path ]]; then
+    core_found=true
+  fi
+done
+
+# Copy the executable to the main temporary directory of the test so we can analyze it later.  Also
+# keep the libraries around, because gdb will search for them in their exact original locations.
+if "$core_found"; then
+  keep_library_dir=1
+  (
+    set -x
+    cp "$SAMPLES_DIR"/sample "$TEST_TMPDIR"
+  )
+fi
+
+exit "$exit_code"
